Make virtual trackball center on pan gesture start point
Instead of having a fixed (invisible) virtual trackball and letting
the user drag on that, move the virtual trackball to center on
wherever the user touches.  That way only the relative change in
position matters---if the user swipes horizontally the model will
rotate around the y axis, even if the user performs that horizontal
swipe down near the bottom of the screen.  Previously that would have
twisted the model in a funny way because the user would be working
on the very bottom of the virtual trackball.