Add arcball and zoom to the OS X model viewer program (Fixes #3871)
Adds arcball dragging control to the OS X model viewer program,
adds QuaternionGetEuler function.  When the user drags the model
with the mouse we use QuaternionGetEuler to get the new pitch
roll and yaw, and we send those to the CModelDocument.  That way
the text fields stay up to date too.

Adds zoom control via mousewheel actions.  Enforces min scale of
0.2 and max scale of 5.0 for mousewheel zooming.